Move all dev-only scripts into scripts/ganache
